+++ backend/lib/tenantee/tenant/tenant.ex
@@ -4,6 +4,7 @@ defmodule Tenantee.Tenant do
   """
 
   alias Tenantee.Tenant.Schema
+  alias Tenantee.Tenant.Communication.Schema, as: CommunicationSchema
   alias Tenantee.Repo
   import Ecto.Query
 
@@ -21,7 +22,7 @@ defmodule Tenantee.Tenant do
   """
   def get_tenant_by_id(id) do
     with tenant <- Repo.get(Schema, id) do
-      if tenant, do: {:ok, tenant}, else: {:error, :not_found}
+      if tenant, do: {:ok, tenant |> Repo.preload(:communications)}, else: {:error, :not_found}
     end
   end
 
@@ -29,7 +30,7 @@ defmodule Tenantee.Tenant do
   Gets a list of all tenants.
   """
   def get_all_tenants do
-    Repo.all(Schema)
+    Repo.all(Schema) |> Repo.preload(:communications)
   end
 
   @doc """
@@ -39,7 +40,7 @@ defmodule Tenantee.Tenant do
     with {:ok, tenant} <- get_tenant_by_id(id),
          changeset <- Schema.changeset(tenant, attrs),
          {:ok, updated_tenant} <- Repo.update(changeset) do
-      {:ok, updated_tenant}
+      {:ok, updated_tenant |> Repo.preload(:communications)}
     else
       {:error, :not_found} ->
         {:error, :not_found}
@@ -54,6 +55,41 @@ defmodule Tenantee.Tenant do
            from(t in Schema, where: t.id == ^id)
            |> Repo.delete_all() do
       if affected_rows > 0, do: {:ok, :deleted}, else: {:error, :not_found}
+    end
+  end
+
+  @doc """
+  Creates a new communication for a tenant.
+  """
+  def add_communication(id, type, value) do
+    with {:ok, tenant} <- get_tenant_by_id(id),
+         {:ok, communication} <-
+           %CommunicationSchema{}
+           |> CommunicationSchema.changeset(%{
+             type: type,
+             value: value,
+             tenant: tenant |> Map.from_struct()
+           })
+           |> Repo.insert() do
+      Schema.add_communication(tenant, communication)
+      |> Repo.update()
+    else
+      {:error, error} -> {:error, error}
+    end
+  end
+
+  @doc """
+  Deletes an existing communication for a tenant.
+  """
+  def remove_communication(tenant_id, communication_id) do
+    with {:ok, tenant} <- get_tenant_by_id(tenant_id),
+         communication when not is_nil(communication) <-
+           Repo.get(CommunicationSchema, communication_id) do
+      Schema.remove_communication(tenant, communication)
+      |> Repo.update()
+    else
+      {:error, error} -> {:error, error}
+      nil -> {:error, :not_found}
     end
   end
 end